chore: Moved `Azure Functions` to **V4**
