--- Stats/function.json
@@ -1,17 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "function",
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req",
-      "route": "Stats/{system?}",
-      "methods": ["post", "get"]
-    },
-    {
-      "type": "http",
-      "direction": "out",
-      "name": "$return"
-    }
-  ]
-}
