+++ package.json
@@ -2,12 +2,14 @@
   "name": "azf-statistics",
   "version": "2.1.3",
   "description": "",
+  "main": "src/functions/*.js",
   "scripts": {
     "start": "func start",
     "test": "biome check && node --test tests/**/*.test.js",
     "lint:fix": "biome check --write"
   },
   "dependencies": {
+    "@azure/functions": "^4.0.0",
     "@vestfoldfylke/loglady": "^1.5.3",
     "mongodb": "^7.1.1"
   },
