+++ src/lib/create-statistic.js
